Move test to Tests/TestOutputChange
