+++ Tests/TestFuncs/TestBasicFuncs/FuncBasicTests.swift
@@ -1,5 +1,4 @@
 import Foundation
-import XCTest
 
 class BasicFuncTests: MockoloTestCase {
 
@@ -48,25 +47,6 @@ class BasicFuncTests: MockoloTestCase {
         verify(srcContent: simpleFuncs,
                dstContent: simpleMockFuncMock,
                useTemplateFunc: true)
-    }
-
-    func testFileWriteBehavior() throws {
-        verify(srcContent: simpleFuncs,
-               dstContent: simpleFuncsMock)
-
-        let expectedAttributes = try FileManager.default.attributesOfItem(atPath: dstFilePath)
-        let expectedCreationDate = (expectedAttributes[.creationDate] as! Date).timeIntervalSinceReferenceDate
-        let expectedModificationDate = (expectedAttributes[.modificationDate] as! Date).timeIntervalSinceReferenceDate
-
-        verify(srcContent: simpleFuncs,
-               dstContent: simpleFuncsMock)
-
-        let attributes = try FileManager.default.attributesOfItem(atPath: dstFilePath)
-        let creationDate = (attributes[.creationDate] as! Date).timeIntervalSinceReferenceDate
-        let modificationDate = (attributes[.modificationDate] as! Date).timeIntervalSinceReferenceDate
-
-        XCTAssertEqual(creationDate, expectedCreationDate)
-        XCTAssertEqual(modificationDate, expectedModificationDate)
     }
 }
